Average the lag power in float64
The FFT autocorrelation averages each row's power spectrum with
np.mean(axis=0). On single-precision input that is a float32 running sum
that adds one row at a time, whereas the per-lag loop it replaced took its
outer mean over a contiguous 1-D array, which numpy sums pairwise. The
drift grows with the row count: on 400k rows of a correlated complex64
series tau moved 7.6e-5 relative to double precision, against 3.0e-7 for
the old loop. At production row counts that moves the printed N_eff in
its fourth or fifth digit and can flip the last digit of sigma.

Accumulate in float64. The same series then lands 3.9e-8 from double
precision, and the cost is two lag-length accumulators.

+++ tabascal/tab_tools.py
@@ -203,10 +203,12 @@ def _integrated_autocorr_time(arr: np.ndarray, axis: int) -> float:
     # Zero padding gives linear rather than circular autocovariance. Only its
     # real part is used, so sum the powers of separate real/imaginary rFFTs.
     # Average in frequency space before inversion to keep just one lag vector.
+    # Accumulate that average in float64: across ~1e6 rows a float32 sum drifts
+    # by ~1e-4, which would show in the printed N_eff.
     n_fft = 1 << (2 * n - 1).bit_length()
-    power = np.mean(np.abs(np.fft.rfft(m.real, n=n_fft)) ** 2, axis=0)
+    power = np.mean(np.abs(np.fft.rfft(m.real, n=n_fft)) ** 2, axis=0, dtype=np.float64)
     if np.iscomplexobj(m):
-        power += np.mean(np.abs(np.fft.rfft(m.imag, n=n_fft)) ** 2, axis=0)
+        power += np.mean(np.abs(np.fft.rfft(m.imag, n=n_fft)) ** 2, axis=0, dtype=np.float64)
     covariance = np.fft.irfft(power, n=n_fft)[:n]
     tau = 1.0
     for k in range(1, n):
